+++ www/script/interface.php
@@ -1,6 +1,6 @@
-<?php 
+<?php
 
-require __DIR__ .'/../config.php'; 
+require __DIR__ .'/../config.php';
 
 
 // Initialize technical object to manage hooks of page. Note that conf->hooks_modules contains array of hook context
@@ -12,7 +12,10 @@ $parameters=array();
 $reshook=$hookmanager->executeHooks('doActionInterface',$parameters,$context, $context->action);    // Note that $action and $object may have been modified by hook
 if ($reshook < 0) $context->setEventMessages($hookmanager->error,$hookmanager->errors,'errors');
 
-if (!empty($reshook))
+
+if (empty($reshook))
 {
-    
+	if($context->action == 'productimg'){
+		outputProductImg(GETPOST('p','int'), GETPOST('f','aZ'));
+	}
 }